+++ valhalla/meili/routing.h
@@ -4,7 +4,6 @@
 #include <valhalla/baldr/double_bucket_queue.h>
 #include <valhalla/baldr/graphid.h>
 #include <valhalla/baldr/graphreader.h>
-#include <valhalla/baldr/pathlocation.h>
 #include <valhalla/midgard/distanceapproximator.h>
 #include <valhalla/midgard/pointll.h>
 #include <valhalla/sif/costconstants.h>
@@ -289,7 +288,7 @@ using labelset_ptr_t = std::shared_ptr<LabelSet>;
  */
 std::unordered_map<uint16_t, uint32_t>
 find_shortest_path(baldr::GraphReader& reader,
-                   const std::vector<baldr::PathLocation>& destinations,
+                   const std::vector<Location>& destinations,
                    uint16_t origin_idx,
                    labelset_ptr_t labelset,
                    const midgard::DistanceApproximator<midgard::PointLL>& approximator,